gateway/requirements.txt
fastapi
uvicorn
httpx
slowapi
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
